[Rule Registry] Rewrite APM registry rules for Observability (#117740)
* [Actionable Observability] Rewrite APM registry rules for Observability

* removing apm's rule registry test

* Moving everything under observability_api_integration

* removing old observability directory under api_integrations since it's not being used

* removing observability from api_integration tests

* renaming file

* moving test to x-pack/test/rule_registry

* Adding error handing to cleanupTargetIndices
