--- x-pack/test/api_integration/apis/observability/index.ts
@@ -1,14 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { FtrProviderContext } from '../../ftr_provider_context';
-
-export default function observabilityApiIntegrationTests({ loadTestFile }: FtrProviderContext) {
-  describe('Observability specs', () => {
-    loadTestFile(require.resolve('./annotations'));
-  });
-}
