+++ x-pack/test/rule_registry/spaces_only/tests/basic/index.ts
@@ -22,9 +22,7 @@ export default ({ loadTestFile, getService }: FtrProviderContext): void => {
       await deleteSpaces(getService);
     });
 
-    // Trial
-    loadTestFile(require.resolve('./get_alert_by_id'));
-    loadTestFile(require.resolve('./update_alert'));
-    loadTestFile(require.resolve('./create_rule'));
+    // Basic
+    loadTestFile(require.resolve('./bootstrap'));
   });
 };
